// src/helpers/types.ts
export interface SortType {
  /**
   * The column name.
   * This is the same as `header.prop` from `TableColumnInfo` interface.
   */
  prop: string;
  /** The sort order. */
  order: 'asc' | 'desc';
}
// ...
export type TableGroup<T = any> = {
  type: 'group';
  title: string;
  /** An array of nested header children, which can be yet another group column. */
  headerChildren: TableColumn<T>[];
};

/**
 * The data structure of a column, which will be used for the `headers`
 * prop of the `DatatableWrapper` props, as well as the `TableHeader` props.
 */
export type TableColumnInfo<T = any> = {
  type?: 'column';
  /**
   * The prop name for the header. This prop should exist in the table body's data
   * so that the column can render a non-empty cell. Moreover, each header should
   * have unique `prop` field.
   */
  prop: keyof T;
  /** The title for the header. */
  title?: string;
  /** Custom render the table header cell. */
  headerCell?: (icon: ReactNode, sortedProp: SortType) => ReactNode;
  /**
   * Custom render the table body cell. This is a function with the row data as parameter.
   */
  cell?: (row: T) => ReactNode;
  /** The props passed to the table headers under `tbody`. */
  thProps?: ThHTMLAttributes<HTMLTableCellElement>;
  // TODO(imballinst): update `cellProps` to return entire props like `thProps` above.
  /** The props passed to the table columns under `tbody`. */
  cellProps?: {
    /**
     * Set a custom `className` to the cell. This can be a raw string or
     * a function with the row data as parameter.
     */
    className?: string | ((row: T) => string);
    /**
     * Set a custom `style` to the cell. This can be a raw object or
     * a function with the row data as parameter.
     */
    style?: CSSProperties | ((row: T) => CSSProperties);
  };
  /**
   * Determines whether the column should be filterable or not.
   * When no headers include `isFilterable: true`, then the filter control
   * will not be rendered.
   */
  isFilterable?: boolean;
  /**
   * Determines whether the column is sortable or not.
   */
  isSortable?: boolean;
  /**
   * Determines whether the column is a checkbox or not. When set, then
   * the column will be a checkbox, both the headers and the rest of the rows.
   */
  checkbox?: {
    /** The prop of the other headers that should be used as the unique ID. */
    idProp: string;
    /** The custom `className` for the checkbox input. */
    className?: string;
  };
  /**
   * Determines the cell alignment. This is a "shortcut" for aligning items instead
   * of having to use `cellProps.className` or `cellProps.style` to achieve
   * the same thing.
   */
  alignment?: {
    // TODO(imballinst): consider if we need vertical alignment as well or not.
    /** Sets the horizontal alignment of the cell. */
    horizontal?: 'left' | 'right' | 'center';
  };
};

/**
 * The type for the table column, which can be either a group or an actual column.
 */
export type TableColumn<T = any> = TableGroup<T> | TableColumnInfo<T>;